+++ backend/src/main/java/com/ssafy/s309/domain/user/dto/SettingsUpdateRequest.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 
 public record SettingsUpdateRequest(
     String name,
-    Integer age,
+    Short age,
     String gender,
     String phone,
     BigDecimal height,
@@ -14,4 +14,4 @@ public record SettingsUpdateRequest(
     Boolean isMedicated,
     BigDecimal targetLow,
     BigDecimal targetHigh,
-    Integer weekStartDay) {}
+    Short weekStartDay) {}
